migrate create_instance spec
